+++ docs/_coverpage.md
@@ -1,0 +1,13 @@
+<!-- ![logo](media/f5_white_24x24.svg) -->
+![logo](media/f5.png ':size=40%')
+
+# F5 VSCode Extension <small>v3.0-pending</small>
+
+> Accelerate your F5 work flows
+
+- REST based
+- VSCode extensible
+- Single workspace
+
+[GitHub](https://github.com/f5devcentral/vscode-f5)
+[Get Started](#the-f5-vscode-extension)
