+++ docs/index.html
@@ -9,23 +9,17 @@
   <meta name="description" content="Accelerate your F5 automation with this vsode extension">
   <meta name="viewport"
     content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+
     <!-- Stylesheets -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/docsify-themeable@0/dist/css/theme-simple.css" title="Simple">
+    <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify-themeable@0/dist/css/theme-simple.css" title="Simple">
     <link rel="stylesheet" href="assets/main.css">
-
-    <!-- Alternate Stylesheets -->
-    <!-- <link rel="stylesheet alternate" href="https://cdn.jsdelivr.net/npm/docsify-themeable@0/dist/css/theme-defaults.css" title="Defaults"> -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/docsify-themeable@0/dist/css/theme-simple-dark.css" title="Simple Dark">
-
-  <!-- <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify-themeable@0/dist/css/theme-simple.css">
-  <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify-themeable@0/dist/css/theme-simple-dark.css"> -->
-
-  <!-- <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify/themes/vue.css"> -->
-  <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify/themes/buble.css" title="Buble">
-  <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify/themes/dark.css" title="Dark">
-  <!-- <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify/themes/pure.css"> -->
-
-  <!-- <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify/lib/themes/vue.css"> -->
+    <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify-themeable@0/dist/css/theme-simple-dark.css" title="Simple Dark">
+    <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify/themes/buble.css" title="Buble">
+    <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify/themes/dark.css" title="Dark">
+    <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify/themes/pure.css" title="Pure">
+    
+    <!-- banner sylesheet - testing -->
+    <!-- <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/docsify-top-banner-plugin@latest/dist/style.css"> -->
 </head>
 
 <body>
@@ -35,12 +29,11 @@
       alias: {
         '/changelog': 'https://raw.githubusercontent.com/f5devcentral/vscode-f5/corkscrew/CHANGELOG.md'
       },
-      topMargin: 190, // default: 0
+      topMargin: 90, // default: 0
       name: 'F5 VSCode Docs',
       repo: 'https://github.com/f5devcentral/vscode-f5',
       homepage: 'README.md',
       auto2top: true,
-      // basePath: '/docs/',
       coverpage: true,
       executeScript: true,
       loadSidebar: true,
@@ -48,16 +41,15 @@
       mergeNavbar: true,
       maxLevel: 2,
       subMaxLevel: 2,
-      formatUpdated: '{MM}/{DD} {HH}:{mm}',
-      formatUpdated: function (time) {
-        // ...
+      // formatUpdated: '{MM}/{DD} {HH}:{mm}',
+      // formatUpdated: function (time) {
+      //   // ...
 
-        return time;
-      },
+      //   return time;
+      // },
 
       // themeColor: '#DCDCDC'
-      // logo: 'media/f5.png'
-      // search: 'auto',
+
       search: {
         // maxAge: 86400000, // Expiration time, the default one day
         maxAge: 5, // Expiration time, the default one day
@@ -68,6 +60,7 @@
         depth: 6,
         hideOtherSidebarContent: false, // whether or not to hide other sidebar content
       },
+      // banner config - testihng
       // https://github.com/Plugin-contrib/docsify-plugin/tree/master/packages/docsify-top-banner-plugin
       // topBanner: {
       //   content: 'yay - banner!'
@@ -96,7 +89,6 @@
   <script src="//cdn.jsdelivr.net/npm/docsify/lib/docsify.min.js"></script>
   <script src="//cdn.jsdelivr.net/npm/docsify-copy-code"></script>
   <script src="//cdn.jsdelivr.net/npm/docsify/lib/plugins/search.min.js"></script>
-  <!-- <script src="//unpkg.com/prismjs/components/prism-bash.min.js"></script> -->
   
   <script src="//cdn.jsdelivr.net/npm/prismjs@1/components/prism-bash.min.js"></script>
   <script src="//cdn.jsdelivr.net/npm/prismjs@1/components/prism-markdown.min.js"></script>
@@ -105,12 +97,12 @@
   <script src="//cdn.jsdelivr.net/npm/prismjs@1/components/prism-json.js"></script>
   <script src="//cdn.jsdelivr.net/npm/prismjs@1/components/prism-jq.js"></script>
   <script src="//cdn.jsdelivr.net/npm/prismjs@1/components/prism-tcl.js"></script>
-  <script src="//cdn.jsdelivr.net/npm/docsify-top-banner-plugin@latest/dist/index.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/docsify@4/lib/plugins/external-script.min.js"></script>
-  <script src="//cdn.jsdelivr.net/npm/docsify-themeable@0"></script>
+  
+  <!-- banner plugin - testing -->
+  <!-- <script src="//cdn.jsdelivr.net/npm/docsify-top-banner-plugin@latest/dist/index.js"></script> -->
 
-  <!-- docsify-themeable (latest v0.x.x) -->
-  <!-- <script src="https://cdn.jsdelivr.net/npm/docsify-themeable@0/dist/js/docsify-themeable.min.js"></script> -->
+  <script src="//cdn.jsdelivr.net/npm/docsify@4/lib/plugins/external-script.min.js"></script>
+  <script src="https://cdn.jsdelivr.net/npm/docsify-themeable@0/dist/js/docsify-themeable.min.js"></script>
 
 </body>
